wip: refactor menu
